requirements.txt
gradio
pandas
langchain
langchain-core
langchain-google-genai
langchain-community
langchain-huggingface
faiss-cpu
google-generativeai
sentence-transformers
requests
google-search-results
python-dotenv
beautifulsoup4
fastapi
uvicorn
pydantic